Seperate Application data from the Review form and create new component called ApplicationPreview
